--- a/src/openms/include/OpenMS/ANALYSIS/XLMS/OPXLSpectrumProcessingAlgorithms.h
+++ b/src/openms/include/OpenMS/ANALYSIS/XLMS/OPXLSpectrumProcessingAlgorithms.h
@@ -38,6 +38,7 @@
 #include <OpenMS/KERNEL/StandardTypes.h>
 #include <OpenMS/KERNEL/Peak1D.h>
 #include <OpenMS/METADATA/DataArrays.h>
+#include <OpenMS/CHEMISTRY/SimpleTSGXLMS.h>
 #include <numeric>
 #include <vector>
 
@@ -95,6 +96,26 @@ namespace OpenMS
             const DataArrays::IntegerDataArray& exp_charges,
             DataArrays::FloatDataArray& ppm_error_array,
             double intensity_cutoff = 0.0);
+
+            /**
+             * @brief Computes a spectrum alignment while considering fragment charges. Uses TSGXLMS::SimplePeak for the theoretical spectrum and its charges. Does not consider intensities.
+             * @param alignment The empty alignment, that will be filled by the algorithm
+             * @param fragment_mass_tolerance The peak mass tolerance
+             * @param fragment_mass_tolerance_unit_ppm True if the given tolerance is a ppm tolerance, false if tolerance is in Da
+             * @param theo_spectrum The first spectrum to be aligned (preferably the theoretical one)
+             * @param exp_spectrum the second spectrum to be aligned (preferably the experimental one)
+             * @param theo_charges IntegerDataArray with charges for the theo_spectrum
+             * @param exp_charges IntegerDataArray with charges for the exp_spectrum
+            * @param ppm_error_array empty FloatDataArray to be filled with per peak ppm errors
+             * @param intensity_cutoff Peaks will only be aligned if intensity1 / intensity2 > intensity_cutoff, with intensity1 being the lower of the two compared peaks and intensity2 the higher one. Set to 0 to ignore intensity differences.
+             */
+            static void getSpectrumAlignmentSimple(
+                  std::vector<std::pair<Size, Size> > & alignment,
+                  double fragment_mass_tolerance,
+                  bool fragment_mass_tolerance_unit_ppm,
+                  const std::vector< SimpleTSGXLMS::SimplePeak >& theo_spectrum,
+                  const PeakSpectrum& exp_spectrum,
+                  const DataArrays::IntegerDataArray& exp_charges);
 
       /**
        * @brief Deisotopes a spectrum and stores the determined charges in an IntegerDataArray