--- a/src/openms/include/OpenMS/CHEMISTRY/Residue.h
+++ b/src/openms/include/OpenMS/CHEMISTRY/Residue.h
@@ -390,6 +390,9 @@ public:
     bool isInResidueSet(const String& residue_set);
     //@}
 
+    /// helper for mapping residue types to letters for Text annotations and labels
+    static char residueTypeToIonLetter_(ResidueType res_type);
+
     /// ostream iterator to write the residue to a stream
     friend OPENMS_DLLAPI std::ostream& operator<<(std::ostream& os, const Residue& residue);