--- a/src/openms/include/OpenMS/CHEMISTRY/sources.cmake
+++ b/src/openms/include/OpenMS/CHEMISTRY/sources.cmake
@@ -28,6 +28,7 @@ RNaseDigestion.h
 TheoreticalSpectrumGenerator.h
 SpectrumAnnotator.h
 TheoreticalSpectrumGeneratorXLMS.h
+SimpleTSGXLMS.h
 SvmTheoreticalSpectrumGenerator.h
 SvmTheoreticalSpectrumGeneratorSet.h
 SvmTheoreticalSpectrumGeneratorTrainer.h