I started with a data set, that took about one hour to analyze on two CPU cores on my laptop and tried to make the tool as fast as possible without changing the output in any way.
Most of the changes concern the generation of theoretical spectra, but I timed the entire tool. 

Starting conditions:
`OpenPepXLLF took 53:40 m (wall), 01:47:03 h (CPU), 2.48 s (system), 01:47:01 h (user).`

I previously calculated many metrics for every candidate. I changed the main loop to only compute the stuff really necessary for the main score that is used for ranking and filtering the hits and removed String annotations from the theoretical spectra. For this I added a new parameter for adding charges to `TheoreticalSpectrumGeneratorXLMS`, because the charges were included in the parameter for metainfo annotations. The old full workflow is now only applied to the top X hits for each spectrum (X being an input parameter of the tool).
New time:
`OpenPepXLLF took 26:50 m (wall), 53:27 m (CPU), 1.72 s (system), 53:25 m (user).`

I implemented a simpler TSG called SimpleTSGXLMS (Simple TheoreticalSpectrumGenerator for CrossLinking MS). For a start I copied the old TSGXLMS and removed any operations and parameters concerning String annotations. Not much improvement was expected here, since they were turned off anyway, but some String operations and declarations were done in the old code even with annotations turned off.
`OpenPepXLLF took 22:17 m (wall), 44:28 m (CPU), 0.90 s (system), 44:27 m (user).`

The generation of neutral losses was determined to be the most inefficient part of the tool.
I simplified the generation of losses by precomputing a database of the losses I use for each residue.
`OpenPepXLLF took 06:51 m (wall), 13:40 m (CPU), 0.64 s (system), 13:40 m (user).`

Changed the SimpleTSGXLMS to use a vector of SimplePeak structs containing only an mz value and a charge as theoretical spectra. Sorting of spectra is done using a comparator only comparing the mz values. Not as big of an improvement, but mainly the sorting of spectra got more efficient, probably because there is no need to keep track of a separate DataArray of charges and sort it in parallel to the spectrum itself.
`OpenPepXLLF took 06:15 m (wall), 12:29 m (CPU), 1.05 s (system), 12:28 m (user).`

Further simplified the generation of neutral losses. Instead of using sets to keep losses for each peptide fragment unique (I know, was never a good idea in the first place, but was still 10x faster than the old code in the original TheoreticalSpectrumGenerator at the time)
now I precomputed the masses for the two losses I actually use (H20 and NH3) and use a struct of two booleans for each residue or peptide fragment to keep track of when I need to apply each neutral loss.
`OpenPepXLLF took 03:09 m (wall), 06:18 m (CPU), 0.93 s (system), 06:17 m (user).`

In the `Residue` class the full residue mass is precomputed for each residue, but getting the monoisotopic weight using `getMonoWeight()` for each fragment type used stuff like `EmpiricalFormula("H2O")` and calculating the weights from chemical compositions for each call.
I defined constants in the `Residue` class that get computed once, making the `getMonoWeight()`
function of `Residue` a simple addition for each fragment type. This should speed up every code we have, that uses this function without any changes in the results, I did not have to update any tests for this.
Final time:
`OpenPepXLLF took 02:33 m (wall), 05:06 m (CPU), 1.23 s (system), 05:05 m (user).`

So I reduced the runtime of the entire tool by a factor of almost 25.

@timosachsenberg we should implement something similar for the losses in the TheoreticalSpectrumGenerator, but maybe in a more general way if we need more types of neutral losses there.