@enetz 2018-12-07
Added a filter that cancels analysis of a candidate if it does not have at least 2 matched linear peaks per peptide. This avoids the generation of more complex theoretical spectra and additional alignments for matches that I would not trust after manual validation of the spectra anyway. This of course does change the results. It reduces the number of completely useless hits in the output and therefore the output file size. The runtime improvement from this will probably be more variable depending on the dataset.
New time:
`OpenPepXLLF took 01:46 m (wall), 03:33 m (CPU), 0.44 s (system), 03:32 m (user).`