@timosachsenberg 2018-12-10
That's a really nice speedup. Would it be possible to introduce some templated methods that can deal with both simple and standard spectra?
Have you also tried if you get some speedup if you use two vectors (one for mz and one for charge) instead of a vector of simplepeaks?