@enetz 2018-12-10
@timosachsenberg I have not tried using separate vectors for mz and charge, because that would make much of the code more complex again. The point was to avoid having a complex sorting algorithm that has to sort multiple vectors by the values in one of them. The sorting is still one of the slowest parts of the tool. Maybe I will try an experiment to see how much faster it is to sort a `vector < double >` than a `vector < SimplePeak >`, but I am not planning to make that change in my tool anytime soon.
With the current struct I don't think it makes sense to make templated methods. It would be possible if I added functions like `getMZ()` to make the interface more similar, but the only function in OpenPepXL where that would work would be `XQuestScores::matchOddsScore(...)`. I duplicated the alignment algorithm to make it simpler by removing everything intensity related and the collection of mass deviations for each match. But I need the intensity checks for aligning two experimental spectra, so I can't remove the old version either. These are the only two functions where I use SimplePeaks.