@enetz 2018-12-10
@timosachsenberg for completeness' sake I did an artificial experiment on sorting `double` and `SimplePeak` vectors. Sorting 50k vectors of length 2k that start off in the reverse order:
`double`: 16 sek
`SimplePeak`: 22 sek

There is a speedup of about 25% for the sorting when using simple `double` vectors. In OpenPepXL, sorting spectra takes about 30% of the total runtime now, so I could reduce the runtime by 25% of 30%, so about 7.5% total, but would have to add the overhead of a more complex sorting to keep two vectors synchronized.