@enetz 2018-12-10
The test files were not deleted, they are just much smaller now. That is because of the last optimization step, where I stop analyzing candidates that don't have at least 2 linear ion matches per peptide. Previously the top 5 hits for each spectrum were always reported, no matter how bad they were. Now those that don't pass this low bar are skipped for the more complicated analysis steps and therefore also don't get reported.
I need to implement some kind of quality filter for the reporting of results and this is a first step and a very low bar in my opinion.