@hroest 2018-12-12
@enetz are any of your improvements applicable to the regular TheoSpectrumGenerator? Eg the precomputation of certain losses -- could that also speed up the regular class? If not, can you add a note / warning to the docu of the TheoSpectrumGenerator that the class is slow and that developers should look at your implementation on how to make it fast?