@enetz 2018-12-12
@hroest my implementation of neutral losses is very limited and only considers H2O and NH3 losses. But the method should be extensible to any number of additional losses if needed. I just don't know which losses, if any, other people use with the regular TSG. We should certainly implement something similar there. Also the changes to Residue::getMonoWeight() should lead to an improvement for the regular TSG already, because it uses that function often as well.

I will add some documentation.